move messaging to chrome-events
