+++ shared/js/background/background.es6.js
@@ -22,14 +22,6 @@ const settings = require('./settings.es6')
 const tabManager = require('./tab-manager.es6')
 const redirect = require('./redirect.es6')
 
-// popup will ask for the browser type then it is created
-chrome.runtime.onMessage.addListener((req, sender, res) => {
-    if (req.getBrowser) {
-        res(utils.getBrowserName());
-    }
-    return true;
-});
-
 function Background() {
   // clearing last search on browser startup
   settings.updateSetting('last_search', '')
